force open ccta with ctrl+shift+o
